Responsive a frontpage
