+++ src/views/Home/Css/Benefit.css
@@ -70,19 +70,58 @@ body {
   gap: 10px;
 }
 
-@media screen and (max-width: 768px) {
-  .Benefit-text {
-    /* display: none; */
-  }
+/* Responsive */
 
+/* 
+@media screen and (max-width: 768px) {
+  .container,
+  .title-2,
+  .title-3 {
+    display: none;
+  }
+  .title-1 {
+    font-size: 30px;
+    line-height: 30px;
+  }
+  .title {
+    font-size: 9px;
+    line-height: 9px;
+  }
+  .Benefit-text {
+    font-size: 22px;
+    line-height: 22px;
+    width: auto;
+    height: auto;
+    padding: 10px;
+  }
   .Rectangle-Benfit {
-    /* display: none; */
+    width: auto;
+    height: auto;
+    margin: 10px;
+    padding: 10px;
   }
 }
 
-@media screen and (max-width: 480px) {
+@media screen and (max-width: 575px) {
+  .title-1 {
+    font-size: 30px;
+    line-height: 30px;
+  }
+  .title {
+    font-size: 12px;
+    line-height: 12px;
+  }
   .Benefit-text {
-    font-size: 20px;
-    line-height: 18px;
+    font-size: 24px;
+    line-height: 28px;
+  }
+  .Rectangle-Benfit {
+    grid-template-columns: 1fr;
+  }
+} */
+
+@media only screen and (max-width: 768px) {
+  .container-2 {
+    display: none;
   }
 }
