+++ src/views/Home/style.css
@@ -15,19 +15,21 @@ body {
   padding: auto;
 }
 
-.header.sticky {
-  position: fixed;
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-}
 .header {
-  height: 101px;
-  position: relative;
+  width: 100%;
   background: linear-gradient(
     to left,
-    var(--col-1),
-    var(--col-purplel),
-    var(--col-2)
+    var(--col-light-gray),
+    var(--col-light-purple),
+    var(--col-off-white)
   );
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: sticky;
+  top: 0;
+  z-index: 1000;
 }
 
 .logo {
@@ -82,7 +84,8 @@ body {
   width: 100px;
   height: 48px;
 }
-.btn1, .btn2 {
+.btn1,
+.btn2 {
   width: 100px;
   height: 48px;
   flex-shrink: 0;
@@ -91,89 +94,168 @@ body {
   color: white;
   margin-right: 5px;
   box-shadow: 0px 10px 20px 0px rgba(0, 0, 0, 0.15);
-  transition: color 0.3s; 
+  transition: color 0.3s;
 }
 
+.btn1:hover,
+.btn2:hover {
+  color: grey;
+}
+@import url("https://fonts.googleapis.com/css2?family=Outfit:wght@400;600&display=swap");
 
-.btn1:hover, .btn2:hover {
+:root {
+  --col-light-gray: #eaeaea;
+  --col-off-white: #f5f5f5;
+  --col-light-blue: #babdf7;
+  --col-dark-blue: #c8dcff;
+  --col-light-purple: #e8c6ff;
+  --col-dark-purple: #f2edf9;
+  --col-mint-green: #cdfff7;
+}
+
+body {
+  margin: 0;
+  padding: 0;
+  font-family: "Outfit", sans-serif;
+}
+
+.header {
+  width: 100%;
+  background: linear-gradient(
+    to left,
+    var(--col-light-gray),
+    var(--col-light-purple),
+    var(--col-off-white)
+  );
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.logo img {
+  max-width: 100px;
+  max-height: 100px;
+  width: auto;
+  height: auto;
+}
+
+.tab {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.navigation {
+  list-style: none;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0;
+  margin: 0;
+}
+
+.navigation-item a {
+  margin: 0 20px;
+  text-decoration: none;
+  color: rgba(94, 94, 94, 0.71);
+  font-size: 14px;
+  font-weight: 600;
+  line-height: 35px;
+  letter-spacing: 3.5px;
+  transition: color 0.1s ease-in-out;
+}
+
+.navigation-item a:hover {
+  color: rgba(18, 17, 17, 0.71);
+}
+
+.button-header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+}
+
+.button-primary,
+.button-secondary {
+  border-radius: 20px;
+  background: #787eff;
+  color: white;
+  padding: 12px 24px;
+  box-shadow: 0px 10px 20px 0px rgba(0, 0, 0, 0.15);
+  transition: background-color 0.3s, color 0.3s;
+  cursor: pointer;
+}
+
+.button-primary:hover,
+.button-secondary:hover {
+  background-color: #6a6fe8; /* Adjust the background color on hover */
   color: grey;
 }
 
-/* Media Queries */
 @media only screen and (max-width: 768px) {
   .header {
-    height: 80px;
-  }
-
-  .logo {
-    width: 80px;
-    height: 80px;
+    height: auto;
+    /* width: auto; */
   }
 
   .logo img {
-    width: 80px;
-    height: 80px;
+    max-width: 60px;
+    max-height: 60px;
   }
 
-  .home li a {
+  .navigation-item a {
     font-size: 12px;
     line-height: 25px;
     letter-spacing: 2.5px;
-  }
-
-  .Btn-Header {
-    margin-right: 10px;
-    width: 80px;
-    height: 40px;
-  }
-
-  .btn1 {
-    width: 80px;
-    height: 40px;
-    margin-right: 3px;
-  }
-
-  .btn2 {
-    width: 80px;
-    height: 40px;
+    padding: 10px;
   }
 }
-
-@media only screen and (max-width: 480px) {
+@media only screen and (max-width: 575px) {
   .header {
-    height: 60px;
-  }
-
-  .logo {
-    width: 60px;
-    height: 60px;
+    flex-direction: column;
+    flex-wrap: wrap; /* Allows items to wrap */
+    padding: 10px 0;
   }
 
   .logo img {
-    width: 60px;
-    height: 60px;
+    width: 100%; /* Ensures the image is responsive */
+    height: auto;
+    max-width: 60px;
+    max-height: 60px;
   }
 
-  .home li a {
-    font-size: 10px;
-    line-height: 20px;
-    letter-spacing: 2px;
+  .navigation {
+    flex-direction: column;
+    width: 100%; /* Ensures navigation takes full width */
   }
 
-  .Btn-Header {
-    margin-right: 5px;
-    width: 60px;
-    height: 30px;
+  .navigation-item a {
+    font-size: 14px; /* Minimum font size */
+    letter-spacing: 1px;
+    margin: 10px 0;
   }
 
-  .btn1 {
-    width: 60px;
-    height: 30px;
-    margin-right: 2px;
+  .button-header {
+    flex-direction: column; /* Stack buttons vertically */
+    width: 100%; /* Ensures buttons take full width */
+    margin-top: 10px;
   }
 
-  .btn2 {
-    width: 60px;
-    height: 30px;
+  .button-primary,
+  .button-secondary {
+    width: auto; /* Allows buttons to grow */
+    min-width: 60px; /* Minimum button width */
+    padding: 10px 20px;
   }
 }
